+++ src/index.tsx
@@ -9,6 +9,7 @@ import { I18nextProvider } from 'react-i18next';
 import { ThemeProvider } from 'styled-components';
 
 import App from 'components/App/App';
+import ApiContextProvider from 'contexts/ApiContext/ApiContext';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -17,7 +18,9 @@ root.render(
   <React.StrictMode>
     <I18nextProvider i18n={i18n}>
       <ThemeProvider theme={theme}>
-        <App />
+        <ApiContextProvider>
+          <App />
+        </ApiContextProvider>
       </ThemeProvider>
     </I18nextProvider>
   </React.StrictMode>