+++ src/types/User.type.ts
@@ -1,0 +1,14 @@
+/**
+ * User Schema declaration
+ * @example
+ * {
+ *  discordId: '275631064574197761',
+ *  displayName: 'Jordina',
+ *  creationTime: '2019-12-29T03:39:07'
+ * }
+ */
+export type User = {
+  discordId: string;
+  displayName: string;
+  creationTime: string;
+};